minor: moved some functions from card_issuer main.py to some helper.py file. -@iamserda
